Move libengine namespace to engine namespace
