--- test/libengine/BUILD
@@ -1,29 +1,0 @@
-#
-# Copyright 2020 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      https://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-load("@rules_cc//cc:defs.bzl", "cc_test")
-
-package(default_testonly = True)
-
-cc_test(
-    name = "engine_bind_test",
-    srcs = ["engine_bind_test.cc"],
-    deps = [
-        "//src/libengine:engine_bind",
-        "@com_google_googletest//:gtest_main",
-    ],
-    size = "small",
-)
